Read description if you want
Fixed:
Collision stuff in PhysicsEngine, PhysicsObject, and Player that were
causing new bugs after the earlier changes to controls. Started working
on collision with doors. It works on some of them, but is buggy on other
doors. Will examine more tonight to see if it's due to my code or if
it's due to the actual door objects themselves...