Fixed, mostly. Read description.
Hitbox.X and Hitbox.Y return Coordinates.X and Coordinates.Y
respectively. But those are the location relative to the screen. The
should return Position.X and Position.Y, which are the location relative
to the room. That was the big problem I was running into with the rooms.
Now I gotta figure out why I need to stand on top of right-facing doors
in order to interact with them...