Our first Mummy...
...is in the lobby. I had to add code/text to GameManager.Update,
Room.Draw, and Lobby.txt to get his guy up and running. Yes, he's
probably too spry for a mummy at this point, but I just wanted to test
out different ways to make different kinds of enemies move and also give
a little example for anybody that's interested in trying to make a
badguy. It's really not that hard -- just have to choose the correct
velocities (or accelerations and velocity limits, if you wanna mess with
those) based on the bad guy's stimuli
When I wake up I will add onto the Enemy class and make some more bad
guys.