New Enemy movement
Cut and pasted the Mummy AI to the default case in Update, then gave the
Mummy new behavior to test. Just walks around, and turns around whenever
it bumps into another enemy, wall, or object, or when it is about to
step over a ledge. Let me know what you think. I'm gonna find a way to
make a bad guy that crawls up and down walls/floors/ceilings...