Here's how I envisioned sequence of the BGM during death...
As soon as the HP reaches 0, all music stops. All you can hear is the
dying character grunting each time an enemy attacks his dying body, as
well as the impact with the ground whenever he hits the ground. Then the
game over music starts playing at the same time the Death Screen pops
up.
This is how it is right here, and I urge everyone to experience this
hilariousness firsthand. If enough people disagree and thinks that the
game over music start playing as soon as HP reaches 0, while he's
getting beat up, then we can change it back. But seriously, I think most
will agree that this is funny.